Move the lexer and object definitions to separate files
