+++ src/object.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include <string>
+#include <variant>
+
+namespace anzu {
+
+using object = std::variant<int, bool>;
+
+bool is_literal(const std::string& token);
+anzu::object parse_literal(const std::string& token);
+
+}
